Java 25, Spring Boot 4.0.3, Spring 7.0.5 Upgrade, Event Driven Route Refresh, Client Access Control Changes and Bug Fixes

+++ api-gateway/src/main/java/org/eclipse/ecsp/gateway/ratelimit/GatewayRateLimiter.java
@@ -101,7 +101,7 @@ public class GatewayRateLimiter extends RedisRateLimiter {
         }
 
         int replenishRate = routeConfig.getReplenishRate();
-        int burstCapacity = routeConfig.getBurstCapacity();
+        long burstCapacity = routeConfig.getBurstCapacity();
         int requestedTokens = routeConfig.getRequestedTokens();
 
         LOGGER.debug("Rate limit for route {}: replenishRate={}, burstCapacity={}, requestedTokens={}, namespace={}",
